Move LM out
