+++ fish_speech_core/lib/lib.rs
@@ -1,4 +1,4 @@
 pub mod audio;
 pub mod codec;
 pub mod config;
-pub mod models;
+pub mod lm;
