+++ fish_speech_core/lib/lm/mod.rs
@@ -1,6 +1,6 @@
 pub mod dual_ar;
 pub mod generate;
-pub mod ops;
+mod ops;
 pub mod sampling;
 pub mod utils;
