--- fish_speech_core/lib/models/mod.rs
@@ -1,1 +1,0 @@
-pub mod lm;
