+++ server/lib/handlers/send_hidden_states.rs
@@ -5,7 +5,7 @@ use anyhow::Context;
 use axum::body::Body;
 use axum::{extract::State, http::StatusCode, response::Response, Json};
 use candle_core::Tensor;
-use fish_speech_core::models::lm::utils::{encode::encode_chunks, text::preprocess_text};
+use fish_speech_core::lm::utils::{encode::encode_chunks, text::preprocess_text};
 use serde::Deserialize;
 use std::io::{Cursor, Write};
 use std::sync::Arc;
