+++ server/lib/utils/mod.rs
@@ -2,7 +2,7 @@ pub mod load;
 
 use candle_core::{Device, Tensor};
 use fish_speech_core::config::WhichLM;
-use fish_speech_core::models::lm::utils::encode::{load_prompt_text, PromptEncoder};
+use fish_speech_core::lm::utils::encode::{load_prompt_text, PromptEncoder};
 use serde::Deserialize;
 use std::collections::HashMap;
 use std::path::Path;
